requirements.txt
Django==1.8.4
PyJWT==1.4.0
SQLAlchemy==1.0.12
argparse==1.2.1
dj-database-url==0.4.0
django-postgrespool==0.3.0
djangorestframework==3.3.3
httplib2==0.9.1
oauth2client==1.5.0
oauthlib==1.0.3
psycopg2==2.6.1
pyasn1==0.1.8
pyasn1-modules==0.0.7
python-openid==2.2.5
python-social-auth==0.2.14
requests==2.9.1
requests-oauthlib==0.6.1
rsa==3.2
six==1.9.0
whitenoise==3.0
wsgiref==0.1.2
